tech debt(ml inference): abstract simulate pipeline logic
Moved the simulate pipeline and get document from the add inference
logic to a test pipeline logic dedicated to that component. This
significantly cleans up the ml_inference_logic.
