+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/search_index/pipelines/ml_inference/__mocks__/ml_inference_logic.mock.ts
@@ -1,0 +1,23 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export const mockMlInferenceValues: any = {
+  addInferencePipelineModal: {
+    configuration: {
+      destinationField: '',
+      modelID: '',
+      pipelineName: '',
+      sourceField: '',
+    },
+    indexName: '',
+    step: 0,
+  },
+  mlInferncePipeline: undefined,
+};
+jest.mock('../ml_inference_logic', () => ({
+  MLInferenceLogic: () => mockMlInferenceValues,
+}));
