+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/search_index/pipelines/ml_inference/ml_inference_logic.ts
@@ -8,7 +8,6 @@
 import { kea, MakeLogicType } from 'kea';
 
 import { IndicesGetMappingIndexMappingRecord } from '@elastic/elasticsearch/lib/api/types';
-import { IngestSimulateResponse } from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
 
 import {
   formatPipelineName,
@@ -16,17 +15,12 @@ import {
   getMlModelTypesForModelConfig,
   parseMlInferenceParametersFromPipeline,
 } from '../../../../../../../common/ml_inference_pipeline';
-import { Status, HttpError } from '../../../../../../../common/types/api';
+import { Status } from '../../../../../../../common/types/api';
 import { MlInferencePipeline } from '../../../../../../../common/types/pipelines';
 import { Actions } from '../../../../../shared/api_logic/create_api_logic';
 
 import { getErrorsFromHttpResponse } from '../../../../../shared/flash_messages/handle_api_errors';
 
-import {
-  GetDocumentsApiLogic,
-  GetDocumentsArgs,
-  GetDocumentsResponse,
-} from '../../../../api/documents/get_document_logic';
 import {
   CachedFetchIndexApiLogic,
   CachedFetchIndexApiLogicValues,
@@ -60,16 +54,6 @@ import {
   FetchMlInferencePipelinesArgs,
   FetchMlInferencePipelinesResponse,
 } from '../../../../api/pipelines/fetch_ml_inference_pipelines';
-import {
-  SimulateExistingMlInterfacePipelineApiLogic,
-  SimulateExistingMlInterfacePipelineArgs,
-  SimulateExistingMlInterfacePipelineResponse,
-} from '../../../../api/pipelines/simulate_existing_ml_inference_pipeline';
-import {
-  SimulateMlInterfacePipelineApiLogic,
-  SimulateMlInterfacePipelineArgs,
-  SimulateMlInterfacePipelineResponse,
-} from '../../../../api/pipelines/simulate_ml_inference_pipeline_processors';
 
 import { isConnectorIndex } from '../../../../utils/indices';
 import {
@@ -78,7 +62,11 @@ import {
   sortSourceFields,
 } from '../../../shared/ml_inference/utils';
 
-import { AddInferencePipelineFormErrors, InferencePipelineConfiguration } from './types';
+import {
+  AddInferencePipelineFormErrors,
+  AddInferencePipelineSteps,
+  InferencePipelineConfiguration,
+} from './types';
 
 import {
   validateInferencePipelineConfiguration,
@@ -92,12 +80,6 @@ export const EMPTY_PIPELINE_CONFIGURATION: InferencePipelineConfiguration = {
   pipelineName: '',
   sourceField: '',
 };
-
-export enum AddInferencePipelineSteps {
-  Configuration,
-  Test,
-  Review,
-}
 
 const API_REQUEST_COMPLETE_STATUSES = [Status.SUCCESS, Status.ERROR];
 const DEFAULT_CONNECTOR_FIELDS = ['body', 'title', 'id', 'type', 'url'];
@@ -131,8 +113,6 @@ interface MLInferenceProcessorsActions {
     CreateMlInferencePipelineResponse
   >['apiSuccess'];
   createPipeline: () => void;
-  getDocumentApiError: Actions<GetDocumentsArgs, GetDocumentsResponse>['apiError'];
-  getDocumentApiSuccess: Actions<GetDocumentsArgs, GetDocumentsResponse>['apiSuccess'];
   makeAttachPipelineRequest: Actions<
     AttachMlInferencePipelineApiLogicArgs,
     AttachMlInferencePipelineResponse
@@ -141,20 +121,11 @@ interface MLInferenceProcessorsActions {
     CreateMlInferencePipelineApiLogicArgs,
     CreateMlInferencePipelineResponse
   >['makeRequest'];
-  makeGetDocumentRequest: Actions<GetDocumentsArgs, GetDocumentsResponse>['makeRequest'];
   makeMLModelsRequest: TrainedModelsApiLogicActions['makeRequest'];
   makeMappingRequest: Actions<GetMappingsArgs, GetMappingsResponse>['makeRequest'];
   makeMlInferencePipelinesRequest: Actions<
     FetchMlInferencePipelinesArgs,
     FetchMlInferencePipelinesResponse
-  >['makeRequest'];
-  makeSimulateExistingPipelineRequest: Actions<
-    SimulateExistingMlInterfacePipelineArgs,
-    SimulateExistingMlInterfacePipelineResponse
-  >['makeRequest'];
-  makeSimulatePipelineRequest: Actions<
-    SimulateMlInterfacePipelineArgs,
-    SimulateMlInterfacePipelineResponse
   >['makeRequest'];
   mappingsApiError: Actions<GetMappingsArgs, GetMappingsResponse>['apiError'];
   mlModelsApiError: TrainedModelsApiLogicActions['apiError'];
@@ -168,41 +139,11 @@ interface MLInferenceProcessorsActions {
   setInferencePipelineConfiguration: (configuration: InferencePipelineConfiguration) => {
     configuration: InferencePipelineConfiguration;
   };
-  setPipelineSimulateBody: (simulateBody: string) => {
-    simulateBody: string;
-  };
-  setSimulatePipelineErrors(errors: string[]): { errors: string[] };
-  simulateExistingPipelineApiError: Actions<
-    SimulateExistingMlInterfacePipelineArgs,
-    SimulateExistingMlInterfacePipelineResponse
-  >['apiError'];
-  simulateExistingPipelineApiReset: Actions<
-    SimulateExistingMlInterfacePipelineArgs,
-    SimulateExistingMlInterfacePipelineResponse
-  >['apiReset'];
-  simulateExistingPipelineApiSuccess: Actions<
-    SimulateExistingMlInterfacePipelineArgs,
-    SimulateExistingMlInterfacePipelineResponse
-  >['apiSuccess'];
-  simulatePipeline: () => void;
-  simulatePipelineApiError: Actions<
-    SimulateMlInterfacePipelineArgs,
-    SimulateMlInterfacePipelineResponse
-  >['apiError'];
-  simulatePipelineApiReset: Actions<
-    SimulateMlInterfacePipelineArgs,
-    SimulateMlInterfacePipelineResponse
-  >['apiReset'];
-  simulatePipelineApiSuccess: Actions<
-    SimulateMlInterfacePipelineArgs,
-    SimulateMlInterfacePipelineResponse
-  >['apiSuccess'];
 }
 
 export interface AddInferencePipelineModal {
   configuration: InferencePipelineConfiguration;
   indexName: string;
-  simulateBody: string;
   step: AddInferencePipelineSteps;
 }
 
@@ -211,12 +152,7 @@ export interface MLInferenceProcessorsValues {
   createErrors: string[];
   existingInferencePipelines: MLInferencePipelineOption[];
   formErrors: AddInferencePipelineFormErrors;
-  getDocumentApiErrorMessage: HttpError | undefined;
-  getDocumentApiStatus: Status;
-  getDocumentData: typeof GetDocumentsApiLogic.values.data;
-  getDocumentsErr: string;
   index: CachedFetchIndexApiLogicValues['indexData'];
-  isGetDocumentsLoading: boolean;
   isLoading: boolean;
   isPipelineDataValid: boolean;
   mappingData: typeof MappingsApiLogic.values.data;
@@ -226,13 +162,6 @@ export interface MLInferenceProcessorsValues {
   mlInferencePipelinesData: FetchMlInferencePipelinesResponse | undefined;
   mlModelsData: TrainedModel[] | null;
   mlModelsStatus: Status;
-  showGetDocumentErrors: boolean;
-  simulateExistingPipelineData: typeof SimulateExistingMlInterfacePipelineApiLogic.values.data;
-  simulateExistingPipelineStatus: Status;
-  simulatePipelineData: typeof SimulateMlInterfacePipelineApiLogic.values.data;
-  simulatePipelineErrors: string[];
-  simulatePipelineResult: IngestSimulateResponse | undefined;
-  simulatePipelineStatus: Status;
   sourceFields: string[] | undefined;
   supportedMLModels: TrainedModel[];
 }
@@ -251,11 +180,6 @@ export const MLInferenceLogic = kea<
     setInferencePipelineConfiguration: (configuration: InferencePipelineConfiguration) => ({
       configuration,
     }),
-    setPipelineSimulateBody: (simulateBody: string) => ({
-      simulateBody,
-    }),
-    setSimulatePipelineErrors: (errors: string[]) => ({ errors }),
-    simulatePipeline: true,
   },
   connect: {
     actions: [
@@ -265,20 +189,6 @@ export const MLInferenceLogic = kea<
       ['makeRequest as makeMappingRequest', 'apiError as mappingsApiError'],
       TrainedModelsApiLogic,
       ['makeRequest as makeMLModelsRequest', 'apiError as mlModelsApiError'],
-      SimulateExistingMlInterfacePipelineApiLogic,
-      [
-        'makeRequest as makeSimulateExistingPipelineRequest',
-        'apiSuccess as simulateExistingPipelineApiSuccess',
-        'apiError as simulateExistingPipelineApiError',
-        'apiReset as simulateExistingPipelineApiReset',
-      ],
-      SimulateMlInterfacePipelineApiLogic,
-      [
-        'makeRequest as makeSimulatePipelineRequest',
-        'apiSuccess as simulatePipelineApiSuccess',
-        'apiError as simulatePipelineApiError',
-        'apiReset as simulatePipelineApiReset',
-      ],
       CreateMlInferencePipelineApiLogic,
       [
         'apiError as createApiError',
@@ -291,12 +201,6 @@ export const MLInferenceLogic = kea<
         'apiSuccess as attachApiSuccess',
         'makeRequest as makeAttachPipelineRequest',
       ],
-      GetDocumentsApiLogic,
-      [
-        'apiError as getDocumentApiError',
-        'apiSuccess as getDocumentApiSuccess',
-        'makeRequest as makeGetDocumentRequest',
-      ],
     ],
     values: [
       CachedFetchIndexApiLogic,
@@ -307,18 +211,8 @@ export const MLInferenceLogic = kea<
       ['data as mappingData', 'status as mappingStatus'],
       TrainedModelsApiLogic,
       ['data as mlModelsData', 'status as mlModelsStatus'],
-      SimulateExistingMlInterfacePipelineApiLogic,
-      ['data as simulateExistingPipelineData', 'status as simulateExistingPipelineStatus'],
-      SimulateMlInterfacePipelineApiLogic,
-      ['data as simulatePipelineData', 'status as simulatePipelineStatus'],
       FetchMlInferencePipelineProcessorsApiLogic,
       ['data as mlInferencePipelineProcessors'],
-      GetDocumentsApiLogic,
-      [
-        'data as getDocumentData',
-        'status as getDocumentApiStatus',
-        'error as getDocumentApiErrorMessage',
-      ],
     ],
   },
   events: {},
@@ -370,27 +264,6 @@ export const MLInferenceLogic = kea<
       actions.makeMLModelsRequest(undefined);
       actions.makeMappingRequest({ indexName });
     },
-    simulatePipeline: () => {
-      if (values.mlInferencePipeline) {
-        actions.setSimulatePipelineErrors([]);
-        actions.simulateExistingPipelineApiReset();
-        actions.simulatePipelineApiReset();
-        const { configuration } = values.addInferencePipelineModal;
-        if (configuration.existingPipeline) {
-          actions.makeSimulateExistingPipelineRequest({
-            docs: values.addInferencePipelineModal.simulateBody,
-            indexName: values.addInferencePipelineModal.indexName,
-            pipelineName: configuration.pipelineName,
-          });
-        } else {
-          actions.makeSimulatePipelineRequest({
-            docs: values.addInferencePipelineModal.simulateBody,
-            indexName: values.addInferencePipelineModal.indexName,
-            pipeline: values.mlInferencePipeline,
-          });
-        }
-      }
-    },
   }),
   path: ['enterprise_search', 'content', 'pipelines_add_ml_inference_pipeline'],
   reducers: {
@@ -400,25 +273,14 @@ export const MLInferenceLogic = kea<
           ...EMPTY_PIPELINE_CONFIGURATION,
         },
         indexName: '',
-        simulateBody: `[
-
-]`,
         step: AddInferencePipelineSteps.Configuration,
       },
       {
-        getDocumentApiSuccess: (modal, doc) => ({
-          ...modal,
-          simulateBody: JSON.stringify([doc], undefined, 2),
-        }),
         setAddInferencePipelineStep: (modal, { step }) => ({ ...modal, step }),
         setIndexName: (modal, { indexName }) => ({ ...modal, indexName }),
         setInferencePipelineConfiguration: (modal, { configuration }) => ({
           ...modal,
           configuration,
-        }),
-        setPipelineSimulateBody: (modal, { simulateBody }) => ({
-          ...modal,
-          simulateBody,
         }),
       },
     ],
@@ -431,33 +293,12 @@ export const MLInferenceLogic = kea<
         makeCreatePipelineRequest: () => [],
       },
     ],
-    simulatePipelineErrors: [
-      [],
-      {
-        setSimulatePipelineErrors: (_, { errors }) => errors,
-        simulateExistingPipelineApiError: (_, error) => getErrorsFromHttpResponse(error),
-        simulatePipelineApiError: (_, error) => getErrorsFromHttpResponse(error),
-      },
-    ],
   },
   selectors: ({ selectors }) => ({
     formErrors: [
       () => [selectors.addInferencePipelineModal],
       (modal: AddInferencePipelineModal) =>
         validateInferencePipelineConfiguration(modal.configuration),
-    ],
-    getDocumentsErr: [
-      () => [selectors.getDocumentApiErrorMessage],
-      (err: MLInferenceProcessorsValues['getDocumentApiErrorMessage']) => {
-        if (!err) return '';
-        return getErrorsFromHttpResponse(err)[0];
-      },
-    ],
-    isGetDocumentsLoading: [
-      () => [selectors.getDocumentApiStatus],
-      (status) => {
-        return status === Status.LOADING;
-      },
     ],
     isLoading: [
       () => [selectors.mlModelsStatus, selectors.mappingStatus],
@@ -468,12 +309,6 @@ export const MLInferenceLogic = kea<
     isPipelineDataValid: [
       () => [selectors.formErrors],
       (errors: AddInferencePipelineFormErrors) => Object.keys(errors).length === 0,
-    ],
-    showGetDocumentErrors: [
-      () => [selectors.getDocumentApiStatus],
-      (status: MLInferenceProcessorsValues['getDocumentApiStatus']) => {
-        return status === Status.ERROR;
-      },
     ],
     mlInferencePipeline: [
       () => [
@@ -509,24 +344,6 @@ export const MLInferenceLogic = kea<
           pipelineName: configuration.pipelineName,
           sourceField: configuration.sourceField,
         });
-      },
-    ],
-    simulatePipelineResult: [
-      () => [
-        selectors.simulatePipelineStatus,
-        selectors.simulatePipelineData,
-        selectors.simulateExistingPipelineStatus,
-        selectors.simulateExistingPipelineData,
-      ],
-      (
-        status: Status,
-        simulateResult: IngestSimulateResponse | undefined,
-        exStatus: Status,
-        exSimulateResult: IngestSimulateResponse | undefined
-      ) => {
-        if (exStatus === Status.SUCCESS) return exSimulateResult;
-        if (status === Status.SUCCESS) return simulateResult;
-        return undefined;
       },
     ],
     sourceFields: [
