+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/search_index/pipelines/ml_inference/test_pipeline.tsx
@@ -27,7 +27,7 @@ import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n-react';
 import { CodeEditor } from '@kbn/kibana-react-plugin/public';
 
-import { MLInferenceLogic } from './ml_inference_logic';
+import { TestPipelineLogic } from './test_pipeline_logic';
 
 import './add_inference_pipeline_flyout.scss';
 
@@ -36,16 +36,16 @@ export const TestPipeline: React.FC = () => {
     addInferencePipelineModal: {
       configuration: { sourceField },
       indexName,
-      simulateBody,
     },
     getDocumentsErr,
     isGetDocumentsLoading,
     showGetDocumentErrors,
+    simulateBody,
     simulatePipelineResult,
     simulatePipelineErrors,
-  } = useValues(MLInferenceLogic);
+  } = useValues(TestPipelineLogic);
   const { simulatePipeline, setPipelineSimulateBody, makeGetDocumentRequest } =
-    useActions(MLInferenceLogic);
+    useActions(TestPipelineLogic);
 
   const isSmallerViewport = useIsWithinMaxBreakpoint('s');
   const inputRef = useRef<HTMLInputElement>();
