+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/search_index/pipelines/ml_inference/types.ts
@@ -19,3 +19,9 @@ export interface AddInferencePipelineFormErrors {
   pipelineName?: string;
   sourceField?: string;
 }
+
+export enum AddInferencePipelineSteps {
+  Configuration,
+  Test,
+  Review,
+}
